+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/SimpleExoPlayer.java
@@ -1079,18 +1079,6 @@ public class SimpleExoPlayer extends BasePlayer
     player.stop();
   }
 
-  /**
-   * @deprecated Use {@link #stop()} and {@link #clearMediaItems()} (if {@code reset} is true) or
-   *     just {@link #stop()} (if {@code reset} is false). Any player error will be cleared when
-   *     {@link #prepare() re-preparing} the player.
-   */
-  @Deprecated
-  @Override
-  public void stop(boolean reset) {
-    blockUntilConstructorFinished();
-    player.stop(reset);
-  }
-
   @Override
   public void release() {
     blockUntilConstructorFinished();