+++ libraries/test_utils/src/main/java/androidx/media3/test/utils/Action.java
@@ -339,12 +339,10 @@ public abstract class Action {
     }
   }
 
-  /** Calls {@link Player#stop()} or {@link Player#stop(boolean)}. */
+  /** Calls {@link Player#stop()}. */
   public static final class Stop extends Action {
 
     private static final String STOP_ACTION_TAG = "Stop";
-
-    @Nullable private final Boolean reset;
 
     /**
      * Action will call {@link Player#stop()}.
@@ -353,28 +351,12 @@ public abstract class Action {
      */
     public Stop(String tag) {
       super(tag, STOP_ACTION_TAG);
-      this.reset = null;
-    }
-
-    /**
-     * Action will call {@link Player#stop(boolean)}.
-     *
-     * @param tag A tag to use for logging.
-     * @param reset The value to pass to {@link Player#stop(boolean)}.
-     */
-    public Stop(@Size(max = 23) String tag, boolean reset) {
-      super(tag, STOP_ACTION_TAG);
-      this.reset = reset;
     }
 
     @Override
     protected void doActionImpl(
         ExoPlayer player, DefaultTrackSelector trackSelector, @Nullable Surface surface) {
-      if (reset == null) {
-        player.stop();
-      } else {
-        player.stop(reset);
-      }
+      player.stop();
     }
   }