+++ libraries/test_utils/src/main/java/androidx/media3/test/utils/StubPlayer.java
@@ -185,17 +185,6 @@ public class StubPlayer extends BasePlayer {
     throw new UnsupportedOperationException();
   }
 
-  /**
-   * @deprecated Use {@link #stop()} and {@link #clearMediaItems()} (if {@code reset} is true) or
-   *     just {@link #stop()} (if {@code reset} is false). Any player error will be cleared when
-   *     {@link #prepare() re-preparing} the player.
-   */
-  @Deprecated
-  @Override
-  public void stop(boolean reset) {
-    throw new UnsupportedOperationException();
-  }
-
   @Override
   public void release() {
     throw new UnsupportedOperationException();